plot.ssgraph {ssgraph}	R Documentation

Plot function for S3 class "ssgraph"

Description

Visualizes structure of the selected graphs which could be a graph with links for which their estimated posterior probabilities are greater than 0.5 or graph with the highest posterior probability.

Usage

 ## S3 method for class 'ssgraph'
plot( x, cut = 0.5, layout = layout.circle, ... ) 

Arguments

x
An object of S3 class "ssgraph", from function ssgraph.

cut
Threshold for including the links in the selected graph based on the estimated posterior probabilities of the links; See the examples.

layout
The vertex placement algorithm which is according to igraph package.

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

See Also

ssgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
dat <- bdgraph.sim( n = 50, p = 6, vis = TRUE )
  
out_ss <- ssgraph( data = dat )

plot( out_ss )

plot( out_ss, cut = 0.3 )

## End(Not run)

[Package ssgraph version 1.4 Index]