print.ssgraph {ssgraph}	R Documentation

Print function for S3 class "ssgraph"

Description

Prints the information about the selected graph which could be a graph with links for which their estimated posterior probabilities are greater than 0.5 or graph with the highest posterior probability. It provides adjacency matrix, size and posterior probability of the selected graph.

Usage

## S3 method for class 'ssgraph'
print( x, round = 2, ... )

Arguments

x
An object of S3 class "ssgraph", from function ssgraph.

round
A value to round the probabilities to the specified number of decimal places.

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

See Also

ssgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
dat <- bdgraph.sim( n = 50, p = 6, vis = TRUE )
  
out_ss <- ssgraph( data = dat )
   
print( out_ss )

## End(Not run)

[Package ssgraph version 1.4 Index]