ssgraph-package {ssgraph}	R Documentation

Bayesian Graphical Estimation using Spike-and-Slab Priors

Description

The R package ssgraph provides statistical tools for Bayesian estimation in graphical models by using spike-and-slab priors. The package is implemented the recent improvements in the Bayesian graphical models' literature, including Wang (2015). To speed up the computations, the computationally intensive tasks of the package are implemented in C++ in parallel using OpenMP.

How to cite this package

Whenever using this package, please cite as

Mohammadi R. (2018). ssgraph: Bayesian Graphical Estimation using Spike-and-Slab Priors, 
R package version 1.4, https://cran.r-project.org/package=ssgraph

Author(s)

Reza Mohammadi <a.mohammadi@uva.nl>

References

Wang, H. (2015). Scaling it up: Stochastic search structure learning in graphical models, Bayesian Analysis, 10(2):351-377

George, E.I. and McCulloch, R.E. (1993). Variable selection via Gibbs sampling. Journal of the American Statistical Association, 88(423):881-889

Griffin, J.E. and Brown, P.J. (2010) Inference with normal-gamma prior distributions in regression problems. Bayesian Analysis, 5(1):171-188

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Examples

## Not run: 

library( ssgraph )

# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 100, p = 8, size = 10, vis = TRUE )

# Running algorithm based on GGMs
ssgraph.obj <- ssgraph( data = data.sim, iter = 5000 )

summary( ssgraph.obj )

# To compare the result with true graph
compare( data.sim, ssgraph.obj, colnames = c("True graph", "ssgraph"), vis = TRUE )

## End(Not run)

[Package ssgraph version 1.4 Index]