summary.ssgraph {ssgraph}	R Documentation

Summary function for S3 class "ssgraph"

Description

Provides a summary of the results for function ssgraph.

Usage

## S3 method for class 'ssgraph'
summary( object, round = 2, vis = TRUE, ... )

Arguments

object
An object of S3 class "ssgraph", from function ssgraph.

round
A value for rounding all probabilities to the specified number of decimal places.

vis
Visualize the results.

...
System reserved (no specific usage).

Value

best.graph
The adjacency matrix corresponding to the selected graph which has the highest posterior probability.

p_links
An upper triangular matrix corresponding to the posterior probabilities of all possible links.

K_hat
The estimated precision matrix.

Author(s)

Reza Mohammadi

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

See Also

ssgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
dat <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
   
out_ss <- ssgraph( data = dat )
   
summary( out_ss )

## End(Not run)

[Package ssgraph version 1.4 Index]